% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/precall.r
\name{plot_precall}
\alias{plot_precall}
\title{Plot precision-recall curves}
\usage{
plot_precall(labels, scores, print_friendly = FALSE, use_color, use_linetype,
  axis_label_size = 15, legend_title_size = 13, legend_text_size = 10,
  linewidth = 0.7)
}
\arguments{
\item{labels}{A vector of logical values that encodes the POSITIVE/NEGATIVE status of data}

\item{scores}{A matrix or vector of numeric values that specifies the rankings}

\item{print_friendly}{A boolean}

\item{use_color}{A boolean}

\item{use_linetype}{A boolean}

\item{axis_label_size}{A numeric}

\item{legend_title_size}{A numeric}

\item{legend_text_size}{A numeric}

\item{linewidth}{A numeric}
}
\value{
A \code{ggplot} object representing precision-recall curves.
}
\description{
\code{plot_precall} creates a ggplot object that represents precision-recall
curves corresponding to several rankings.
}
\details{
The \code{labels} encode the POSITIVE/NEGATIVE (\code{TRUE}/\code{FALSE}) status of data
while each column of \code{scores} provides scores corresponding to a ranking of those data:
a smaller score means a smaller rank.

A precision-recall curve is plotted for each column of the \code{scores}. The interpolation
between points is done according to the principle of partial acceptance.

\code{print_friendly} enables to produce print-friendly plots.
If set to \code{TRUE}, the function yields a black and white plot (background and lines)
with a different line type for each curve.
\code{use_color} and \code{use_linetype}, if present, specifies if the curved should be colored or
drawn with different line types.
These parameters override the behavior specified by \code{print_friendly}.

\code{axis_label_size}, \code{legend_title_size}, \code{legend_text_size} and \code{linewidth} are
self-explanatory parameters to tune a few aesthetics.
Besides, the function returns a ggplot object which can be modified a posteriori using the ggplot
interface.
}
\examples{
labels = c(TRUE, FALSE, TRUE, FALSE, TRUE, TRUE, FALSE, FALSE) # 8 pieces of data

ranking1 = c(1,1,8,3,7,6,4,5)
ranking2 = c(4,6,3,3,2,5,6,6)
ranking3 = c(6,-1,3,3,2,4,1,5)

scores = cbind(ranking1, ranking2, ranking3)
colnames(scores) = c("method a", "method b", "method c")

# Default plot
plot_precall(labels, scores)

# Black and white plot
plot_precall(labels, scores, print_friendly=TRUE)

# Print-friendly background with colored curves
plot_precall(labels, scores, print_friendly=TRUE, use_color=TRUE, use_linetype=FALSE)
}

