% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/precall.r
\name{precall_interpolation}
\alias{precall_interpolation}
\title{Computes the point coordinates of a precision-recall curve}
\usage{
precall_interpolation(labels, scores, nb_interpolation_points = 30)
}
\arguments{
\item{labels}{A vector of logical values}

\item{scores}{A vector of numeric values}

\item{nb_interpolation_points}{A numeric}
}
\description{
Computes the point coordinates of a precision-recall curve
}

